add Load menu
add auto  target
damage melee
fix bug
add checkbox select scene

+++ I  am not Alone/Assets/Script/Health.cs
@@ -39,7 +39,7 @@ public class Health : MonoBehaviour
 
     }
 
-    public void HelthDamage (int damage)
+    public void HelthDamage (float damage)
     {
